Web/01 - Generateur Super Secure/source/pyproject.toml
[tool.poetry]
name = "chall-1"
version = "0.1.0"
description = ""
authors = ["tristanqtn <tristan.querton@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
